pyproject.toml
[project]
name = "travel-rt-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "XetPy1030",email = "xetpy1030@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "tortoise-orm[asyncpg] (>=1.1.6,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.tortoise]
tortoise_orm = "app.config.db.TORTOISE_ORM"
